Update Stay.py
To sense movement in Python, you can use a sensor such as a PIR (Passive Infrared) sensor. A PIR sensor detects changes in infrared radiation and can be used to detect the movement of people or animals. Once the sensor detects the movement it can trigger an action such as sending a voice command "stay" using text-to-speech(TTS) library like gTTS.
This script uses the RPi.GPIO library to read input from the PIR sensor on pin 18. Inside an infinite loop, it checks the input of the pin and when it detects a high level signal it triggers the TTS library to convert the text "Stay" to speech and saves it as an mp3 file and then it plays it using os.system("mpg321 stay.mp3") which is a command line based media player.